backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
types-Flask-Cors==4.0.0
Python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
toml==0.10.2
types-toml==0.10.2
openai==1.6.1
tiktoken==0.7.0
mypy==1.10.0
ruff==0.5.0
google-generativeai==0.7.2
google-cloud-aiplatform==1.62.0
coverage==7.6.1
